services/quiz_engine/requirements.txt
Flask==3.1.1
google-generativeai==0.4.1
requests==2.32.4